/uploads RW permission fixed.

+++ wp_setup.sh
@@ -1,5 +1,5 @@
 #/bin/bash
-set -x
+set +x
 
 # Check root pertmissions
 [ $(id -u) != 0 ] && echo "Im not root" && exit 1
@@ -11,7 +11,7 @@ WP_USER_PASS=password
 DB_HOST=localhost
 WP_DB=wordpress
 
-apt-get -y -q update
+apt-get -y -qq update
 apt-get -y install php5-gd libssh2-php mc vim
 
 mysql -u"$MYSQL_ROOT_USER" -p"$MYSQL_ROOT_PASS" -e "\
@@ -20,7 +20,7 @@ mysql -u"$MYSQL_ROOT_USER" -p"$MYSQL_ROOT_PASS" -e "\
                                                      CREATE DATABASE $WP_DB ;  \
                                                      FLUSH PRIVILEGES;"
 wget -q http://wordpress.org/latest.tar.gz
-tar xzf latest.tar.gz && echo "OK. Latest WP successfully downloaded and extracted."
+tar xzf latest.tar.gz && echo "==== OK. Latest WP successfully downloaded and extracted."
 cd wordpress || exit 15
 
 wget -O- https://api.wordpress.org/secret-key/1.1/salt/ 2>/dev/null >salt.tmp
@@ -34,14 +34,20 @@ sed -i -e "s/^define('DB_PASSWORD', .*$/define('DB_PASSWORD', \'${WP_USER_PASS}\
 sed -i -e "s/^define('DB_HOST', .*$/define('DB_HOST', \'${DB_HOST}\');/" wp-config.php
 echo;echo =========================================== WP-CONFIG ===============================================================
 cat wp-config.php
-echo;echo =====================================================================================================================
+echo ==========================================================================================================================
+echo
 
-rsync -a --delete-before . /var/www/html/
-rm -fr wordpress
+rsync -a --delete-before . /var/www/html/ 
+cd .. 
+[ -d wordpress ] && rm -fr wordpress latest.tar.gz
 cd /var/www/html || exit 25
 chown -R vagrant:www-data *
 mkdir /var/www/html/wp-content/uploads
+chmod -R g+w /var/www/html/wp-content/uploads
 chown -R :www-data /var/www/html/wp-content/uploads
 sed -i 's~^  DocumentRoot .*$~  DocumentRoot /var/www/html~' /etc/apache2/sites-enabled/000-default.conf 
 service apache2 restart
+echo;echo ================== Running services: ================================================================================
+ss -tlpn
+echo ================= COMPLETED ========================================================================= exit code: $? ======
